Squashed commit of the following
commit 8f936479 
Author: Manuel Campo <manuel.campo@wabiproject.com> 
Date: Thu Aug 25 2022 16:53:28 GMT-0300 (hora estándar de Argentina) 

    Avoid using MainConfiguration for test


commit cbd81202 
Author: Manuel Campo <manuel.campo@wabiproject.com> 
Date: Thu Aug 25 2022 16:32:51 GMT-0300 (hora estándar de Argentina) 

    Merge & refactor


commit 5219e689 
Author: Manuel Campo <manuel.campo@wabiproject.com> 
Date: Thu Aug 25 2022 16:22:58 GMT-0300 (hora estándar de Argentina) 

    Use proper data types

+++ api/src/main/kotlin/domain/services/UpdatePaymentService.kt
@@ -50,9 +50,9 @@ class UpdatePaymentService(
     )
 
     private fun EncData.toPayment(encData: String) = Payment(
-        paymentId = txnRefNo,
-        supplierOrderId = supplierOrderId!!,
-        amount = amount,
+        paymentId = txnRefNo.toLong(),
+        supplierOrderId = supplierOrderId!!.toLong(),
+        amount = amount.toBigDecimal(),
         paymentOption = paymentOption,
         responseCode = responseCode,
         message = message,
